refactor external snapshotter to use csi-lib-utils/rpc
